FactoryGirl Setup and postgresql
Adds factory_girl_rails to the Gemfile so FactoryGirl can be used
in this project. Also creates the factories.rb file and a support file
to allow the syntactic sugar for FactoryGirl to be used when creating
models from factories.

Additionally, I forgot to setup the database originally to use postgresql
so this fixes that mistake by changing out the database.yml file and removing
the sqlite3 database files. Also includes the "pg" gem in the Gemfile.